requirements.txt
# Instale as dependências usando: pip install -r requirements.txt
django==5.0.6
djangorestframework==3.15.1